+++ src/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import logo from './logo.svg';
-import { Counter } from './features/counter/Counter';
-import { TimePeriod } from './features/timeperiod/TimePeriod';
+import logo from '../logo.svg';
+import { Counter } from '../features/counter/Counter';
+import { TimePeriod } from '../features/timePeriod/TimePeriod';
 import './App.css';
 
 function App() {
