+++ src/tests/App.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import store from './app/store';
-import App from './App';
+import store from '../reducers/store';
+import App from '../components/App';
 
 test('renders learn react link', () => {
   const { getByText } = render(
